requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
moviepy>=1.0.3
imageio-ffmpeg>=0.4.8